update docs, fix bug with port parsing
Docs updated to update copyright in LICENSE and to note that
IPv4 addresses may be used for server parameters (specifically
not mentioning IPv6).

While testing IPv6, I did some testing of IPv4 addresses, and found
an error in the port parsing logic.  IPv4 addresses with no port
were being treated as a port themselves - at least the first octet
of the address.  Tests were added to check the fix.